Refactor mapper into individual class mappers
This takes care of the class growing too long and should allow for more targeted tests
